rename payment routes uniformly
