+++ app/webpacker/components/RegistrationsV2/api/payment/get/getStripeConfig.js
@@ -1,9 +1,9 @@
-import { paymentConfigRoute } from '../../../../../lib/requests/routes.js.erb';
+import { paymentConfigUrl } from '../../../../../lib/requests/routes.js.erb';
 import { fetchJsonOrError } from '../../../../../lib/requests/fetchWithAuthenticityToken';
 
 export default async function getStripeConfig(
   competitionId,
   paymentId,
 ) {
-  return fetchJsonOrError(paymentConfigRoute(competitionId, paymentId));
+  return fetchJsonOrError(paymentConfigUrl(competitionId, paymentId));
 }
